+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Dependency to monocle optics library.
 - Support for minting group tokens.
 - Support for minting series tokens.
+- Added full documentation in docusaurus. Available [here](https://topl.github.io/brambl-cli).
 
 ### Changed